requirements.txt
# LEX - Lender Eligibility X
# Python Dependencies

# Core Framework
streamlit>=1.28.0

# Google Gemini AI
google-generativeai>=0.3.0

# Environment Variables Management
python-dotenv>=1.0.0

# Type Checking (Optional - for development)
typing-extensions>=4.8.0

